voice_to_image/requirements.txt
streamlit==1.41.1
openai==1.61.0
streamlit-audiorec==0.1.3
python-dotenv==1.0.1
Pillow